extract cloudinary logic to separate files
